Move shared messages to api module
