+++ api/src/main/java/api/commands/SelectAllCommand.java
