+++ api/src/main/java/api/configuration/AkkaConfigLoader.java
@@ -3,11 +3,7 @@ package api.configuration;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
 import com.typesafe.config.ConfigValueFactory;
-import lombok.Builder;
-import lombok.Data;
 
-@Data
-@Builder
 public class AkkaConfigLoader {
 
     private static final String AKKA_HOSTNAME_FIELD = "akka.remote.netty.tcp.hostname";
