+++ api/src/main/java/api/messages/QueryErrorMsg.java
@@ -1,6 +1,4 @@
-package store.messages.query;
-
-import api.messages.QueryMetaInfo;
+package api.messages;
 
 public final class QueryErrorMsg extends QueryResponseMsg {
