+++ api/src/main/java/api/messages/QueryMetaInfo.java
