+++ store/build.gradle
@@ -1,3 +1,5 @@
 dependencies {
+    compile project(':api')
+
     testCompile project(':client')
 }
