+++ store/src/main/java/store/actors/Table.java
@@ -14,8 +14,8 @@ import store.messages.partition.SplitPartitionMsg;
 import store.messages.partition.SplitSuccessMsg;
 import store.messages.query.InsertRowMsg;
 import store.messages.query.PartialQueryResultMsg;
-import store.messages.query.QueryErrorMsg;
-import store.messages.query.QuerySuccessMsg;
+import api.messages.QueryErrorMsg;
+import api.messages.QuerySuccessMsg;
 import store.messages.query.SelectAllMsg;
 import store.messages.query.SelectWhereMsg;
 import store.model.BlockedRow;
