+++ store/src/main/java/store/messages/query/CreateTableMsg.java
@@ -1,6 +1,7 @@
 package store.messages.query;
 
 import api.messages.QueryMetaInfo;
+import api.messages.QueryMsg;
 
 import java.util.List;
