+++ store/src/main/java/store/messages/query/InsertRowMsg.java
@@ -1,7 +1,8 @@
 package store.messages.query;
 
 import api.messages.QueryMetaInfo;
-import store.model.Row;
+import api.messages.QueryMsg;
+import api.model.Row;
 
 public class InsertRowMsg extends QueryMsg {
