+++ store/src/test/java/WorkflowTests.java
@@ -7,11 +7,11 @@ import client.config.DatastoreClientModule;
 import com.google.common.collect.ImmutableList;
 import store.configuration.DatastoreModule;
 import store.Datastore;
-import store.messages.query.QueryErrorMsg;
-import store.messages.query.QueryResponseMsg;
-import store.messages.query.QueryResultMsg;
-import store.messages.query.QuerySuccessMsg;
-import store.model.Row;
+import api.messages.QueryErrorMsg;
+import api.messages.QueryResponseMsg;
+import api.messages.QueryResultMsg;
+import api.messages.QuerySuccessMsg;
+import api.model.Row;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
